backend/pyproject.toml
[project]
name = "hypomnemata"
version = "0.1.0"
description = "Local-first repository for captured media and ideas"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "aiosqlite>=0.20",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.17",
    "uuid-utils>=0.10",
    "greenlet>=3.0",
    "pytesseract>=0.3",
    "Pillow>=10.0",
    "pypdf>=4.0",
    "yt-dlp>=2024.1",
    "gallery-dl>=1.31.10",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hypomnemata"]
